Changed the whole unit system
changed the way units work from setting them manualy (without %
working) to setting the main div to the specific units and then getting
all other pixels out from it. Works fine but not if window is resized.
In search of a way to make window resize work. Maybe instead of getting
px transform the data into vw or % ?